### 2024-02-19 #6378 @editorialbot
[PRE REVIEW]: hf_hydrodata: A Python package for accessing hydrologic simulations and observations across the United States
**Submitting author:** @amy-defnet (<a href="http://orcid.org/0000-0003-2261-708X">Amy Defnet</a>)
**Repository:** https://github.com/hydroframe/hf_hydrodata
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
