### 2024-03-01 #6378 @thodson-usgs
Sure, this is right up my alley.
