### 2024-03-07 3ceed3e
fix "climate_response_network" network list (#152)
* fix variable name associated with network list

* increment version
