### 2024-03-10 58086f3
Add noaa dataset with precipitation and temperature of latest week data to support hydrogen emulator (#153)
* add noaa dataset for temp and precip

* update version
