### 2024-03-17 1ccbea4
Add PFB topographic_index entry (#154)
* add pfb topographic_index entry

* fix catalog entry

* reformat

* remove print statement
