### 2024-03-19 78cda7a
Rename conus mapping variables (#156)
* propagate renaming of conus x/y to i/j

* update documentation

* increment version

* fix typo on parameter names
