### 2024-03-22 eb5d277
initial tests for grid_bounds parameter
