### 2024-03-25 225caa7
tests for proper failure conditions
