### 2024-03-25 6b519cc
add tests for get_site_variables function
