### 2024-03-25 b5b6917
Add feature to filter point observations on grid bounds (#157)
* initial commit add conus grid bounds filter

* initial tests for grid_bounds parameter

* tests for proper failure conditions

* add tests for get_site_variables function

* increment version
