+++ src/hf_hydrodata/point.py
@@ -41,6 +41,16 @@ SUPPORTED_FILTERS = [
     "polygon_crs",
     "site_networks",
     "min_num_obs",
+    "grid",
+    "grid_bounds",
+]
+
+# List of SQL tables in the database corresponding to site-type-specific attributes
+SITE_ATTRIBUTE_TABLES = [
+    "streamgauge_attributes",
+    "well_attributes",
+    "snotel_station_attributes",
+    "flux_tower_attributes",
 ]
 
 
@@ -81,6 +91,12 @@ def get_point_data(*args, **kwargs):
         Latitude range bounds for the geographic domain; lesser value is provided first.
     longitude_range : tuple, optional
         Longitude range bounds for the geographic domain; lesser value is provided first.
+    grid : str, optional
+        Value of either 'conus1' or 'conus2'. Used in combination with parameter `grid_bounds`
+        to extract site locations for a specific region of conus coordinates.
+    grid_bounds : list of integers, optional
+        A list of points [left, bottom, right, top] in ij grid coordinates of the grid supplied
+        by the `grid` parameter.
     site_ids : str or list of strings, optional
         Single site ID string or list of desired (string) site identifiers.
     state : str, optional
@@ -265,6 +281,12 @@ def get_point_metadata(*args, **kwargs):
         Latitude range bounds for the geographic domain; lesser value is provided first.
     longitude_range : tuple, optional
         Longitude range bounds for the geographic domain; lesser value is provided first.
+    grid : str, optional
+        Value of either 'conus1' or 'conus2'. Used in combination with parameter `grid_bounds`
+        to extract site locations for a specific region of conus coordinates.
+    grid_bounds : list of integers, optional
+        A list of points [left, bottom, right, top] in ij grid coordinates of the grid supplied
+        by the `grid` parameter.
     site_ids : str or list of strings, optional
         Single site ID string or list of desired (string) site identifiers.
     state : str, optional
@@ -452,7 +474,7 @@ def get_site_variables(*args, **kwargs):
     dataset : str, optional
         Source from which requested data originated. Currently supported: 'usgs_nwis', 'snotel',
         'scan', 'ameriflux'.
-    variable : str, required
+    variable : str, optional
         Description of type of data requested. Currently supported: 'streamflow', 'water_table_depth', 'swe',
         'precipitation', 'air_temp', 'soil_moisture', 'latent_heat', 'sensible_heat',
         'downward_shortwave', 'downward_longwave', 'vapor_pressure_deficit', 'wind_speed'.
@@ -473,6 +495,12 @@ def get_site_variables(*args, **kwargs):
         Latitude range bounds for the geographic domain; lesser value is provided first.
     longitude_range : tuple, optional
         Longitude range bounds for the geographic domain; lesser value is provided first.
+    grid : str, optional
+        Value of either 'conus1' or 'conus2'. Used in combination with parameter `grid_bounds`
+        to extract site locations for a specific region of conus coordinates.
+    grid_bounds : list of integers, optional
+        A list of points [left, bottom, right, top] in ij grid coordinates of the grid supplied
+        by the `grid` parameter.
     site_ids : str or list of strings, optional
         Single site ID string or list of desired (string) site identifiers.
     state : str, optional
@@ -605,6 +633,43 @@ def get_site_variables(*args, **kwargs):
     else:
         lon_query = """"""
 
+    # CONUS grid bounds
+    if "grid_bounds" in options and options["grid_bounds"] is not None:
+        # Make sure that the option "grid" is defined
+        try:
+            assert "grid" in options and options["grid"] in ("conus1", "conus2")
+        except:
+            raise ValueError(
[…]
