### 2024-03-25 7663004
initial commit fix instantaneous wtd filtering
