### 2024-03-26 f00a843
fix test to confirm geographic filtering implemented
