### 2024-03-27 e51b81e
set aggregation to - for instantaneous data, add data catalog entry
