+++ tests/hf_hydrodata/test_point.py
@@ -310,7 +310,7 @@ def test_get_data_wtd_instantaneous():
         dataset="usgs_nwis",
         variable="water_table_depth",
         temporal_resolution="instantaneous",
-        aggregation="instantaneous",
+        aggregation="-",
         date_start="2002-01-01",
         date_end="2002-01-01",
         latitude_range=(30, 35),
@@ -329,7 +329,7 @@ def test_get_metadata_wtd_instantaneous():
         dataset="usgs_nwis",
         variable="water_table_depth",
         temporal_resolution="instantaneous",
-        aggregation="instantaneous",
+        aggregation="-",
         date_start="2002-01-01",
         date_end="2002-01-01",
         latitude_range=(30, 35),
@@ -340,7 +340,7 @@ def test_get_metadata_wtd_instantaneous():
         dataset="usgs_nwis",
         variable="water_table_depth",
         temporal_resolution="instantaneous",
-        aggregation="instantaneous",
+        aggregation="-",
         date_start="2002-01-01",
         date_end="2002-01-01",
         latitude_range=(30, 35),
