### 2024-03-27 dae66dd
exclude from unit test that finds file paths
