### 2024-03-27 98eabea
Fix filtering for instantaneous wtd point observations (#158)
* initial commit fix instantaneous wtd filtering

* fix test to confirm geographic filtering implemented

* feature update for filtering metadata

* add conditional statement

* increment version

* set aggregation to - for instantaneous data, add data catalog entry

* exclude from unit test that finds file paths
