### 2024-03-27 53f4837
Update Arizona copyright text (#159)
* Update Arizona copyright text

* fixed comma in the wrong place
