### 2024-03-28 2f7e2f8
fix bug when omitting temporal_resolution
