### 2024-03-28 e4b82be
fix bug when omitting temporal_resolution (#160)
* fix bug when omitting temporal_resolution

* update version
