### 2024-03-28 #6378 @rwegener2
👋🏻 @jmp75, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
