### 2024-04-16 98b7aa2
Add flow_direction to data catalog and apply new HUC masks
