### 2024-04-16 326a5c4
add nomask=true option and use it in unit test to check flip tiff
