### 2024-04-19 29d0538
add catalog entry for new dataset
