### 2024-04-19 3458aa4
update test to account for added dataset
