### 2024-04-19 2bea890
Add Jasechko et al. 2024 WTD dataset subset (#162)
* update tests to include additional sites caught in filter

* initial commit add jasechko dataset

* incorporate new dataset, add tests

* add data catalog entry

* add dataset description

* reconcile merge conflicts

* add catalog entry for new dataset

* update test to account for added dataset
