+++ tests/hf_hydrodata/test_gridded.py
@@ -314,6 +314,7 @@ def test_files_exist():
             "253",
             "254",
             "522",
+            "525",
         ]:
             paths = gr.get_file_paths(
                 row,
@@ -1217,7 +1218,7 @@ def test_get_datasets():
     """Test get_datasets."""
 
     datasets = hf.get_datasets()
-    assert len(datasets) == 18
+    assert len(datasets) == 19
     assert datasets[0] == "CW3E"
 
     datasets = hf.get_datasets(variable="air_temp")
