### 2024-04-24 0cd3875
initial update to _get_data_sql for fan data
