+++ src/hf_hydrodata/point.py
@@ -68,7 +68,7 @@ def get_point_data(*args, **kwargs):
     ----------
     dataset : str, required
         Source from which requested data originated. Currently supported: 'usgs_nwis', 'snotel',
-        'scan', 'ameriflux', 'jasechko_2024'.
+        'scan', 'ameriflux', 'jasechko_2024', 'fan_2013'.
     variable : str, required
         Description of type of data requested. Currently supported: 'streamflow', 'water_table_depth', 'swe',
         'precipitation', 'air_temp', 'soil_moisture', 'latent_heat', 'sensible_heat',
@@ -242,7 +242,7 @@ def get_point_data(*args, **kwargs):
     if (var_id in (1, 2, 3, 4)) | (var_id in range(6, 25)):
         data_df = _get_data_nc(site_list, var_id, *args, **kwargs)
 
-    elif var_id in (5, 25):
+    elif var_id in (5, 25, 26):
         data_df = _get_data_sql(conn, site_list, var_id, *args, **kwargs)
 
     conn.close()
@@ -2063,14 +2063,18 @@ def _get_data_sql(conn, site_list, var_id, *args, **kwargs):
         Stacked observations data for a single variable, filtered to only sites that
         have the minimum number of observations specified.
     """
-    assert var_id in (5, 25)
+    assert var_id in (5, 25, 26)
     if var_id == 5:
         tbl_name = "wtd_discrete_data"
-        var_names = "w.wtd, w.pumping_status"
+        var_names = "w.site_id, w.date, w.wtd, w.pumping_status"
 
     elif var_id == 25:
         tbl_name = "jasechko_wtd_data"
-        var_names = "w.wtd"
+        var_names = "w.site_id, w.date, w.wtd"
+
+    elif var_id == 26:
+        tbl_name = "fan_wtd_data"
+        var_names = "w.site_id, w.wtd, record_count"
 
     if len(args) > 0 and isinstance(args[0], dict):
         options = args[0]
@@ -2098,47 +2102,91 @@ def _get_data_sql(conn, site_list, var_id, *args, **kwargs):
                 options["date_end"], "%Y-%m-%d"
             ).year
 
-    if ("date_start" not in options) and ("date_end" not in options):
-        date_query = """"""
-        param_list = [min_num_obs]
-    elif ("date_start" not in options) and ("date_end" in options):
-        date_query = """ WHERE w.date <= ?"""
-        param_list = [options["date_end"], min_num_obs, options["date_end"]]
-    elif ("date_start" in options) and ("date_end" not in options):
-        date_query = """ WHERE w.date >= ?"""
-        param_list = [options["date_start"], min_num_obs, options["date_start"]]
-    elif ("date_start" in options) and ("date_end" in options):
-        date_query = """ WHERE w.date >= ? AND w.date <= ?"""
-        param_list = [
-            options["date_start"],
-            options["date_end"],
-            min_num_obs,
-            options["date_start"],
-            options["date_end"],
-        ]
+    # These variables have w.date in the database to be able to filter on
+    if var_id in (5, 25):
+        if ("date_start" not in options) and ("date_end" not in options):
+            date_query = """"""
+            param_list = [min_num_obs]
+        elif ("date_start" not in options) and ("date_end" in options):
+            date_query = """ WHERE w.date <= ?"""
+            param_list = [options["date_end"], min_num_obs, options["date_end"]]
+        elif ("date_start" in options) and ("date_end" not in options):
+            date_query = """ WHERE w.date >= ?"""
+            param_list = [options["date_start"], min_num_obs, options["date_start"]]
+        elif ("date_start" in options) and ("date_end" in options):
+            date_query = """ WHERE w.date >= ? AND w.date <= ?"""
+            param_list = [
[…]
