### 2024-04-25 cc1baf7
remove unused code, add tests
