### 2024-04-25 hydroframe/hf_hydrodata#163 @thodson-usgs
Automate CI/CD
openjournals/joss-reviews#6623

Consider integrating your CI/CD into GitHub using Actions.

Under Github -> Actions -> New workflow
and search for "Publish Python Package" and "Python application". These will publish to pip with a release and automate linting and tests for PRs, respectively.
