### 2024-04-25 hydroframe/hf_hydrodata#164 @thodson-usgs
Add a more complete tutorial
openjournals/joss-reviews#6623

Optional but encouraged. I think users might benefit from having a more developed tutorial example in the docs.
Would it be feasible to use `hf_hydrodata` to set up a "toy" hydrologic model? A template users can run through in a few minutes, which they could riff off in setting up their own workflows.
