### 2024-04-25 f649f05
keep some additional variables on df
