### 2024-04-26 43a5d66
add conus mappings to output dataframe
