### 2024-04-29 26da2cf
Add variables to point.get_site_variables() output (#165)
* keep some additional variables on df

* add conus mappings to output dataframe

* update tests to account for fan dataset in db

* increment version
