### 2024-04-30 1553b10
update README (#166)
* initial commit: address comments

* additional edits
