### 2024-04-30 hydroframe/hf_hydrodata#163 @amy-defnet
Thank you for your comment @thodson-usgs! We do already have an automated CI/CD pipeline implemented, using Jenkins. This pipeline is set up to run when a Pull Request is opened and when any additional commits are made to the PR. Jenkins will run all of our unit tests and display whether they passed or failed back to the GitHub Pull Request. We separately have a Jenkins task that builds and publishes to PyPi. We have added a section to the README to make this more explicit.

Related Pull Request: #166
