### 2024-05-09 #6623 @thodson-usgs
@rwegener2, I've completed my review. The package is functional and documented. However, it mainly provides an API to another database, which I had to register to use, as well as maintain a temporary 2-day PIN. Some question whether it meets the bar of scholarly effort, because a lot of this data is publicly available from other endpoints and Python packages. To that end, it might be good to request either a "state of the field" to better explain its niche, or else a usage example that does something more than query a database or project spatial coordinates.
