### 2024-05-16 #6623 @alessandroamaranto
@rwegener2 My review is still in process. However, I already agree with @thodson-usgs in the state of the field paragraph. It would be beneficial for the authors to elaborate on similar applications, (for example https://cran.r-project.org/web/packages/dataRetrieval/index.html), and highlight how hf_hydrodata offers distinct advantages in these areas.
