### 2024-05-29 eae777e
do not apply coastline mask for conus2
