### 2024-06-04 #6623 @rwegener2
Thanks @amy-defnet for the explanation and for clarifying that information in the manuscript. Please also include a similar description somewhere in your documentation and comment here to let the team know when you've added that.

@thodson-usgs and @alessandroamaranto This seems to be a sufficient statement of scholarly effort. Are you both satisfied?
@alessandroamaranto anything else you need to continue your review?
