### 2024-06-04 #6623 @thodson-usgs
For JOSS, I'm satisfied.
But I'll pick on this point a little more, b/c it isn't the only package in this space.
I don't think you need to list them, but if this package is primarily a preprocessor targeting a particular model (i.e. Parflow) then say so. If it's more general, I would list some of the specific models that it's helpful for. Your readers will appreciate that.
