### 2024-06-05 ea79d4b
add state of the field section to readme
