### 2024-06-05 #6623 @amy-defnet
Thanks @rwegener2: I've updated our [README](https://github.com/hydroframe/hf_hydrodata/blob/main/README.md) to also include the "state of the field" description that we had added to our manuscript.
