### 2024-06-17 #6623 @alessandroamaranto
Sorry for the delay but..my daughter was born! If I remember correctly, I
was basically ok with the idea of the state of the field section for giving
the green light.
I will just have a look at the final version in the afternoon (but just to
cross all the t's), and then proceed.

On Sat, Jun 15, 2024 at 5:34 PM Rachel Wegener ***@***.***>
wrote:

> @alessandroamaranto <https://github.com/alessandroamaranto> Do you have
> any questions to finish your review?
>
