### 2024-07-05 #6623 @rwegener2
Hi @amy-defnet  👋🏻

Just a few small requests for the text:

* line 14: The second comma caused me to stumble. I suggest removing it "in situ observations such as streamflow rate and depth to groundwater ~,~ can be used to ...". Or, alternately, add a comma between "observations" and "such as"
* line 63-65: I was a little confused here - is this sentence is talking about the gridded data products still, or all of the aforementioned data products? Maybe: "Many of these gridded data products, such as inputs and outputs from the national ParFlow model and multiple gridded atmospheric forcing datasets, are not publicly available by other means."
* line 73-75: Is this a complete sentence? Maybe: "However the API is structured to take in compatible input parameters (where applicable) **and** ~to~ make the data querying process as seamless as possible across the different data types." or "However the API is structured to take in compatible input parameters (where applicable) **in order** to make the data querying process as seamless as possible across the different data types."

Cross through is the existing text to be deleted. Bolded text is text to be added. Thanks!
