### 2024-07-09 f4f7402
JOSS paper edits (#170)
