### 2024-07-09 #6623 @amy-defnet
Thanks, @rwegener2! I've adjusted the text given your suggestions.

I created a tagged release, which is version 1.1.12. I archived this with Zenodo and confirmed that the metadata there aligns with the paper title and authors. The DOI is: [10.5281/zenodo.12700800](https://doi.org/10.5281/zenodo.12700800).

Let me know if you need anything else!
