### 2024-07-15 203e6c8
update tests and available substitution keys
