### 2024-07-15 0f57d9c
update tests and allow to substitute calendar year
