+++ tests/hf_hydrodata/test_gridded.py
@@ -1226,7 +1226,7 @@ def test_get_datasets():
     assert datasets[0] == "CW3E"
 
     datasets = hf.get_datasets(grid="conus2")
-    assert len(datasets) == 5
+    assert len(datasets) == 6
     assert datasets[0] == "CW3E"
 
     options = {"variable": "air_temp", "grid": "conus1"}
@@ -1245,7 +1245,7 @@ def test_get_variables():
     assert len(variables) == 8
     assert variables[0] == "air_temp"
     variables = hf.get_variables(grid="conus2")
-    assert len(variables) == 31
+    assert len(variables) == 32
     assert variables[0] == "air_temp"
 
     options = {"dataset": "NLDAS2", "grid": "conus1"}
