### 2024-07-16 c3f372c
Add snow cover dataset (#171)
* add cloud-free snow cover dataset

* fix unsupported characters

* update tests and available substitution keys

* update tests and allow to substitute calendar year

* increment version
