+++ src/hf_hydrodata/gridded.py
@@ -15,12 +15,14 @@ import threading
 import dask
 import requests
 import pyproj
+import rioxarray
+import rasterio
 from dateutil import rrule
 from dateutil.relativedelta import relativedelta
 import numpy as np
 import xarray as xr
 import pandas as pd
-from parflow import read_pfb_sequence, write_pfb
+from parflow import read_pfb_sequence, read_pfb, write_pfb
 from hf_hydrodata.data_model_access import ModelTableRow
 from hf_hydrodata.data_model_access import load_data_model
 from hf_hydrodata.data_catalog import _get_api_headers
@@ -2752,3 +2754,135 @@ class _FileDownloadState:
             while t < self.end_time:
                 self.time_coords.append(t)
                 t = t + relativedelta(months=1)
+
+
+def create_resolution_transform(
+    latitudes: np.ndarray, longitudes: np.ndarray, grid: str
+) -> np.ndarray:
+    """
+    Create transform map from lat-lon degree-based coordinates (EPSG:4326) to conus1 or conus2 grid.
+    This function returns an array transform map that maps the indices of the original grid to the
+    remapped grid (conus1 or conus2).
+
+    Args:
+        latitudes:     A NumPy array of original grid latitude values.
+        longitudes:    A NumPy array of original grid longitude values.
+        grid:          The name of the grid to re-map to: "conus1" or "conus2".
+    Return:
+        A NumPy array containing the transformation indices. The array is the size of the selected
+        conus grid and contains values that represent the row and column grid cell index from the
+        original (unmapped) grid.
+    """
+
+    # Convert lat/lon to coordinate in EPSG:6933
+    transformer = pyproj.Transformer.from_crs("EPSG:4326", "EPSG:6933")
+
+    # For the transform code below, the latitude and longitude array inputs must be the same shape.
+    # It assumes they fully represent the values at each grid cell location.
+    # If instead your file's latitude and longitude fields only provide the *unique* values for each (ndim==1),
+    # the following code uses those to build out the full array.
+    num_lats = latitudes.size
+    num_lons = longitudes.size
+
+    if latitudes.ndim == 1:
+        latitudes = np.repeat(latitudes, num_lons).reshape(num_lats, num_lons)
+
+    if longitudes.ndim == 1:
+        longitudes = np.tile(longitudes, num_lats).reshape(num_lats, num_lons)
+
+    # Transform
+    lons, lats = transformer.transform(latitudes, longitudes)
+
+    # Create dataset with transformed coordinates
+    # Use zeros for the data here since we only care about the coordinate values
+    data_array = xr.DataArray(
+        data=np.zeros(shape=(lats.shape)),
+        dims=["lat", "lon"],
+        coords=[lats[:, 0], lons[0, :]],
+    )
+
+    var_dataset = data_array[:, :]
+    var_dataset = var_dataset.rio.set_spatial_dims(x_dim="lon", y_dim="lat")
+    var_dataset.rio.write_crs("epsg:6933", inplace=True)
+    var_dataset.rio.to_raster("intermediate_mapping.tif")
+
+    ### Create data mapping to CONUS2
+    ### This step outputs a .pfb file that is of CONUS2 size where each value represents the associated
+    ### row and column *index* of the matching MODIS (or other dataset) cell.
+    input_ds = rasterio.open("intermediate_mapping.tif")
+
+    # Extract spatial metadata
+    # The rasterio .transform accessor contains the same information as the gdal .GetGeoTransform(), but
+    # note that the order of returned values of the tuple is different. The following code uses the
[…]
