### 2024-07-23 4e835e9
add conus2 steady state wtd
