### 2024-07-23 b517c03
Add conus2 steady-state water table depth (#172)
* add conus2 steady state wtd

* Change difference from normal to Percenile Rank

---------

Co-authored-by: wh3248 <wh3248@princeton.edu>
