### 2024-07-25 #6623 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
