### 2024-07-26 #6623 @kthyng
Congratulations on your new publication @amy-defnet! Many thanks to @rwegener2 and to reviewers @thodson-usgs and @alessandroamaranto for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
